Unlock All Pets
This works, if you're the type of person who doesn't want to actually collect the pets in battle, but it's fair if you don't want to pay for that stupid membership to "save" the pets. What is best about this hack is how you get to KEEP the pets, even after you log out! :D